Fix all sorts of warnings and stuff.

+++ network/net.cpp
@@ -30,8 +30,8 @@ int send_chat ( int client, const string & body ) {
 
 /* Receive len amount of byte completely (no partial recv) */
 int recv_complete (int sockfd, void *buf, size_t len, int flags) {
-    int bytesRead = 0;
-    int result;
+    size_t bytesRead = 0;
+    ssize_t result;
     while (bytesRead < len) {
         result = recv (sockfd, (char*)buf + bytesRead, len - bytesRead, flags);
         if (result < 1) {
